+++ android/src/main/kotlin/devflowstud/io/device_corner_radius/DeviceCornerRadiusPlugin.kt
@@ -1,35 +1,86 @@
 package devflowstud.io.device_corner_radius
 
-import androidx.annotation.NonNull
-
+import android.app.Activity
+import android.content.Context
+import android.os.Build
+import android.view.RoundedCorner
+import android.view.View
 import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
 
 /** DeviceCornerRadiusPlugin */
-class DeviceCornerRadiusPlugin: FlutterPlugin, MethodCallHandler {
+class DeviceCornerRadiusPlugin: FlutterPlugin, MethodCallHandler, ActivityAware {
   /// The MethodChannel that will the communication between Flutter and native Android
   ///
   /// This local reference serves to register the plugin with the Flutter Engine and unregister it
   /// when the Flutter Engine is detached from the Activity
   private lateinit var channel : MethodChannel
+  private lateinit var context: Context
+  private lateinit var activity: Activity
 
-  override fun onAttachedToEngine(@NonNull flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
+  override fun onAttachedToEngine(flutterPluginBinding: FlutterPlugin.FlutterPluginBinding) {
     channel = MethodChannel(flutterPluginBinding.binaryMessenger, "device_corner_radius")
     channel.setMethodCallHandler(this)
+    context = flutterPluginBinding.applicationContext
   }
 
-  override fun onMethodCall(@NonNull call: MethodCall, @NonNull result: Result) {
-    if (call.method == "getPlatformVersion") {
-      result.success("Android ${android.os.Build.VERSION.RELEASE}")
-    } else {
-      result.notImplemented()
+  override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+    this.activity = binding.activity;
+  }
+
+  override fun onMethodCall(call: MethodCall, result: Result) {
+    when (call.method) {
+      "getPlatformVersion" -> {
+        result.success("Android ${Build.VERSION.RELEASE}")
+      }
+      "getCornerRadius" -> {
+        val view = activity.window.decorView.rootView
+        val cornerRadii = getCornerRadius(view)
+        result.success(cornerRadii)
+      }
+      else -> {
+        result.notImplemented()
+      }
     }
   }
 
-  override fun onDetachedFromEngine(@NonNull binding: FlutterPlugin.FlutterPluginBinding) {
+  private fun getCornerRadius(view: View): Map<String, Any> {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.S) {
+      val insets = view.rootWindowInsets
+      val topLeft = insets?.getRoundedCorner(RoundedCorner.POSITION_TOP_LEFT)?.radius ?: 0f
+      val topRight = insets?.getRoundedCorner(RoundedCorner.POSITION_TOP_RIGHT)?.radius ?: 0f
+      val bottomLeft = insets?.getRoundedCorner(RoundedCorner.POSITION_BOTTOM_LEFT)?.radius ?: 0f
+      val bottomRight = insets?.getRoundedCorner(RoundedCorner.POSITION_BOTTOM_RIGHT)?.radius ?: 0f
+      mapOf(
+        "topLeft" to topLeft,
+        "topRight" to topRight,
+        "bottomLeft" to bottomLeft,
+        "bottomRight" to bottomRight
+      )
+    } else {
+      mapOf(
+        "topLeft" to 0,
+        "topRight" to 0,
+        "bottomLeft" to 0,
+        "bottomRight" to 0
+      )
+    }
+  }
+
+  override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     channel.setMethodCallHandler(null)
   }
+
+  override fun onDetachedFromActivityForConfigChanges() { }
+
+  override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+    this.activity = binding.activity;
+  }
+
+  override fun onDetachedFromActivity() { }
 }