+++ ios/Classes/DeviceCornerRadiusPlugin.swift
@@ -9,6 +9,33 @@ public class DeviceCornerRadiusPlugin: NSObject, FlutterPlugin {
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    result("iOS " + UIDevice.current.systemVersion)
+    if call.method == "getCornerRadius" {
+      let cornerRadius = Float(UIScreen.main.displayCornerRadius)
+      result([
+        "topLeft": cornerRadius,
+        "topRight": cornerRadius,
+        "bottomLeft": cornerRadius,
+        "bottomRight": cornerRadius
+      ])
+    } else {
+      result(FlutterMethodNotImplemented)
+    }
+  }
+}
+
+/// Credits: @kylebshr/ScreenCorners (https://github.com/kylebshr/ScreenCorners)
+/// This code snippet has been taken from the above repository without any modifications.
+extension UIScreen {
+  private static let cornerRadiusKey: String = {
+    let components = ["Radius", "Corner", "display", "_"]
+    return components.reversed().joined()
+  } ()
+
+  public var displayCornerRadius: CGFloat {
+    guard let cornerRadius = self.value(forKey: Self.cornerRadiusKey) as? CGFloat else {
+      assertionFailure("Failed to detect screen corner radius")
+      return 0
+    }
+    return cornerRadius
   }
 }