+++ lib/device_corner_radius_platform_interface.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/rendering.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'device_corner_radius_method_channel.dart';
@@ -23,7 +24,7 @@ abstract class DeviceCornerRadiusPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<String?> getPlatformVersion() {
+  Future<BorderRadius> getCornerRadius() {
     throw UnimplementedError('platformVersion() has not been implemented.');
   }
 }