+++ test/device_corner_radius_method_channel_test.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:device_corner_radius/device_corner_radius_method_channel.dart';
 
 void main() {
-  MethodChannelDeviceCornerRadius platform = MethodChannelDeviceCornerRadius();
   const MethodChannel channel = MethodChannel('device_corner_radius');
 
   TestWidgetsFlutterBinding.ensureInitialized();
@@ -16,9 +14,5 @@ void main() {
 
   tearDown(() {
     channel.setMockMethodCallHandler(null);
-  });
-
-  test('getPlatformVersion', () async {
-    expect(await platform.getPlatformVersion(), '42');
   });
 }